social_media_api/requirements.txt
asgiref==3.8.1
Django==5.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
